requirements.txt
py-sudoku
pulp
